+++ shapez2/__init__.py
@@ -1,4 +1,5 @@
 from . import (
+    gameCode,
     blueprints,
     blueprintsExtraData,
     buildings,