+++ changelog.md
@@ -1,3 +1,12 @@
+## 2.0.0
+- Add gameCode module
+  - Add shape part map generation rareness info to `gameObjects.ShapesConfiguration`
+- Add ability to replace pygamePIL with pygame
+  - Make pygamePIL use submodules like pygame
+- Add support for named colors in feature string rendering
+- Rename `costs` attribute of `research.LinearUpgradeLevel` to `cost`
+- Fix `</color>` having no effect in feature string rendering
+
 ## 1.2.0
 - Raw islands json file changes :
   - Rename `IslandGroups` to `SimilarIslands`